+++ src/components/icons/waiting-icon/index.js
@@ -20,9 +20,10 @@ var WaitingIcon = function WaitingIcon(_ref) {
       title = _ref.title,
       props = (0, _objectWithoutProperties2.default)(_ref, ["color", "title"]);
   return _react.default.createElement("svg", (0, _extends2.default)({
+    xmlns: "http://www.w3.org/2000/svg",
+    viewBox: "0 0 13 3",
     width: "13",
-    height: "3",
-    xmlns: "http://www.w3.org/2000/svg"
+    height: "3"
   }, props), title && _react.default.createElement("title", null, title), _react.default.createElement("path", {
     d: "M5 0h3v3H5V0zM0 0h3v3H0V0zm10 0h3v3h-3V0z",
     fill: color,